🔧 fix(types): JSDoc type casts for array access and branded types
- Add JSDoc type casts to EVM PUSH/SWAP handlers for bytecode array access
- Fix Bytes compare.js files with proper type casts for array indexing
- Add ConstructorInstance interface with encodeParams/decodeParams methods
- Fix Abi.js prototype methods with JSDoc types and this casts
- Replace FFIType.usize with FFIType.u64 in native-loader types
- Fix catch block error types in test files
- Add ItemType[] annotations to ABI test files
- Fix StealthAddress branded type casts
- Fix PublicKey.verify JSDoc to match Secp256k1 verify signature
- Change Array to Array<*> in jsonrpc BatchRequest/Response

Errors reduced: 1177 → 832 (-29%)

+++ src/native-loader/types.ts
@@ -33,7 +33,7 @@ export interface NativeExports {
 
 	// Keccak256 operations
 	primitives_keccak256: {
-		args: [FFIType.ptr, FFIType.usize, FFIType.ptr];
+		args: [FFIType.ptr, FFIType.u64, FFIType.ptr];
 		returns: FFIType.i32;
 	};
 
@@ -57,7 +57,7 @@ export interface NativeExports {
 		returns: FFIType.i32;
 	};
 	primitives_bytes_to_hex: {
-		args: [FFIType.ptr, FFIType.usize, FFIType.ptr, FFIType.ptr];
+		args: [FFIType.ptr, FFIType.u64, FFIType.ptr, FFIType.ptr];
 		returns: FFIType.i32;
 	};
 
@@ -73,7 +73,7 @@ export interface NativeExports {
 
 	// EIP-191 operations
 	primitives_eip191_hash_message: {
-		args: [FFIType.ptr, FFIType.usize, FFIType.ptr];
+		args: [FFIType.ptr, FFIType.u64, FFIType.ptr];
 		returns: FFIType.i32;
 	};
 
@@ -123,21 +123,21 @@ export interface NativeExports {
 
 	// Hash functions
 	primitives_sha256: {
-		args: [FFIType.ptr, FFIType.usize, FFIType.ptr];
+		args: [FFIType.ptr, FFIType.u64, FFIType.ptr];
 		returns: FFIType.i32;
 	};
 	primitives_ripemd160: {
-		args: [FFIType.ptr, FFIType.usize, FFIType.ptr];
+		args: [FFIType.ptr, FFIType.u64, FFIType.ptr];
 		returns: FFIType.i32;
 	};
 	primitives_blake2b: {
-		args: [FFIType.ptr, FFIType.usize, FFIType.ptr];
+		args: [FFIType.ptr, FFIType.u64, FFIType.ptr];
 		returns: FFIType.i32;
 	};
 
 	// RLP operations
 	primitives_rlp_encode_bytes: {
-		args: [FFIType.ptr, FFIType.usize, FFIType.ptr, FFIType.ptr];
+		args: [FFIType.ptr, FFIType.u64, FFIType.ptr, FFIType.ptr];
 		returns: FFIType.i32;
 	};
 	primitives_rlp_encode_uint: {
@@ -145,7 +145,7 @@ export interface NativeExports {
 		returns: FFIType.i32;
 	};
 	primitives_rlp_to_hex: {
-		args: [FFIType.ptr, FFIType.usize, FFIType.ptr, FFIType.ptr];
+		args: [FFIType.ptr, FFIType.u64, FFIType.ptr, FFIType.ptr];
 		returns: FFIType.i32;
 	};
 	primitives_rlp_from_hex: {
@@ -183,11 +183,11 @@ export interface NativeExports {
 
 	// Solidity-compatible hash functions
 	primitives_solidity_keccak256: {
-		args: [FFIType.ptr, FFIType.usize, FFIType.ptr];
+		args: [FFIType.ptr, FFIType.u64, FFIType.ptr];
 		returns: FFIType.i32;
 	};
 	primitives_solidity_sha256: {
-		args: [FFIType.ptr, FFIType.usize, FFIType.ptr];
+		args: [FFIType.ptr, FFIType.u64, FFIType.ptr];
 		returns: FFIType.i32;
 	};
